refactor : [Step2] Exception 위치 및 사용자 에러로 변경
