+++ src/main/java/nextstep/subway/line/ui/LineControllerAdvice.java
@@ -1,5 +1,6 @@
-package nextstep.subway.exception;
+package nextstep.subway.line.ui;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -10,21 +11,26 @@ import nextstep.subway.exception.line.NotFoundSectionException;
 import nextstep.subway.exception.line.NotFoundStationsException;
 
 @RestControllerAdvice
-public class LineControllerAdvice extends ControllerAdvice {
+public class LineControllerAdvice {
 
     @ExceptionHandler(LineAlreadyExistException.class)
     public ResponseEntity<ErrorResponse> lineAlreadyExistException(LineAlreadyExistException e) {
-        return getSeverErrorResponse(e.getMessage());
+        return getBadResponse(e.getMessage());
     }
 
     @ExceptionHandler(NotFoundSectionException.class)
     public ResponseEntity<ErrorResponse> notFoundLineException(NotFoundSectionException e) {
-        return getSeverErrorResponse(e.getMessage());
+        return getBadResponse(e.getMessage());
     }
 
     @ExceptionHandler(NotFoundStationsException.class)
     public ResponseEntity<ErrorResponse> notFoundStationsException(NotFoundStationsException e) {
-        return getSeverErrorResponse(e.getMessage());
+        return getBadResponse(e.getMessage());
+    }
+
+    private ResponseEntity<ErrorResponse> getBadResponse(String message) {
+        ErrorResponse errResponse = new ErrorResponse(HttpStatus.BAD_REQUEST.value(), message);
+        return ResponseEntity.badRequest().body(errResponse);
     }
 
 }
