+++ src/main/java/nextstep/subway/path/ui/PathControllerAdvice.java
@@ -1,6 +1,7 @@
-package nextstep.subway.exception;
+package nextstep.subway.path.ui;
 
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -9,16 +10,21 @@ import nextstep.subway.exception.dto.ErrorResponse;
 import nextstep.subway.exception.path.PathException;
 
 @RestControllerAdvice
-public class PathControllerAdvice extends ControllerAdvice {
+public class PathControllerAdvice {
 
     @ExceptionHandler(DataIntegrityViolationException.class)
     public ResponseEntity<ErrorResponse> noValidatedInputException(DataIntegrityViolationException e) {
-        return getSeverErrorResponse(e.getMessage());
+        return getBadResponse(e.getMessage());
     }
 
     @ExceptionHandler(PathException.class)
     public ResponseEntity<ErrorResponse> invalidDistanceException(PathException e) {
-        return getSeverErrorResponse(e.getMessage());
+        return getBadResponse(e.getMessage());
+    }
+
+    private ResponseEntity<ErrorResponse> getBadResponse(String message) {
+        ErrorResponse errResponse = new ErrorResponse(HttpStatus.BAD_REQUEST.value(), message);
+        return ResponseEntity.badRequest().body(errResponse);
     }
 
 }
