+++ src/main/java/nextstep/subway/station/ui/StationControllerAdvice.java
@@ -1,0 +1,24 @@
+package nextstep.subway.station.ui;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import nextstep.subway.exception.dto.ErrorResponse;
+import nextstep.subway.exception.station.NoStationException;
+
+@RestControllerAdvice
+public class StationControllerAdvice {
+
+    @ExceptionHandler(NoStationException.class)
+    public ResponseEntity<ErrorResponse> stationsAlreadyExistException(NoStationException e) {
+        return getBadResponse(e.getMessage());
+    }
+
+    private ResponseEntity<ErrorResponse> getBadResponse(String message) {
+        ErrorResponse errResponse = new ErrorResponse(HttpStatus.BAD_REQUEST.value(), message);
+        return ResponseEntity.badRequest().body(errResponse);
+    }
+
+}
